Rendered average time and total number of requests
