+++ src/main/js/src/components/Main.tsx
@@ -11,7 +11,6 @@ const Main = (props) => {
   const [broker, setBroker] = useState(null);
   const [topic, setTopic] = useState(null);
   const [isLoaded, setIsLoaded] = useState(false);
-  const [time, setTime] = useState([65, 59, 80, 81, 56, 55, 40]);
 
   const updateBrokerList = () => {
     fetch("/describeBrokers")
@@ -50,33 +49,6 @@ const Main = (props) => {
   useEffect(() => {
     updateList().then(() => setIsLoaded(true));
   }, []);
-
-  const data = {
-    labels: ["January", "February", "March", "April", "May", "June", "July"],
-    datasets: [
-      {
-        label: "My First dataset",
-        fill: false,
-        lineTension: 0.1,
-        backgroundColor: "rgba(75,192,192,0.4)",
-        borderColor: "rgba(75,192,192,1)",
-        borderCapStyle: "butt",
-        borderDash: [],
-        borderDashOffset: 0.0,
-        borderJoinStyle: "miter",
-        pointBorderColor: "rgba(75,192,192,1)",
-        pointBackgroundColor: "#fff",
-        pointBorderWidth: 1,
-        pointHoverRadius: 5,
-        pointHoverBackgroundColor: "rgba(75,192,192,1)",
-        pointHoverBorderColor: "rgba(220,220,220,1)",
-        pointHoverBorderWidth: 2,
-        pointRadius: 1,
-        pointHitRadius: 10,
-        data: time,
-      },
-    ],
-  };
 
   //console.log('NEW RENDER');
   //console.log('brokerdata', broker);
